fix(blocknote): split BlockNote and Workspace, use Y-Sweet Provider, store blocks as Markdown
