+++ package.json
@@ -19,6 +19,7 @@
     "@blocknote/react": "^0.20.0",
     "@blocknote/shadcn": "^0.20.0",
     "@hookform/resolvers": "^3.9.1",
+    "@y-sweet/react": "^0.6.3",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "1.0.4",
